+++ R/model_parameters.R
@@ -12,6 +12,7 @@
 #'  \item{\link[=model_parameters.stanreg]{Bayesian models} (\pkg{rstanarm}, \pkg{brms}, \pkg{MCMCglmm})}
 #'  \item{\link[=model_parameters.principal]{PCA and FA} (\pkg{psych})}
 #'  \item{\link[=model_parameters.lavaan]{CFA and SEM} (\pkg{lavaan})}
+#'   \item{\link[=model_parameters.kmeans]{Cluster models (k-means, ...)}}
 #'  }
 #'
 #' @param model Statistical Model.