+++ R/model_parameters.lavaan.R
@@ -88,8 +88,6 @@ n_parameters.lavaan <- function(x, ...) {
 #' @importFrom insight format_table
 #' @export
 print.parameters_sem <- function(x, ...) {
-  # formatted_table <- parameters_table(x)
-  # cat(insight::format_table(formatted_table))
   .print_model_parms_components(x, pretty_names = TRUE, split_column = "Type")
 }