+++ R/smoothness.R
@@ -15,7 +15,7 @@
 #' @importFrom stats cor sd
 #' @importFrom utils head tail
 #' @export
-check_smoothness <- function(x, method = "cor", lag = 1) {
+smoothness <- function(x, method = "cor", lag = 1) {
   if (lag < 1) {
     lag <- round(lag * length(x))
   }
@@ -31,6 +31,3 @@ check_smoothness <- function(x, method = "cor", lag = 1) {
   smooth
 }
 
-#' @rdname check_smoothness
-#' @export
-smoothness <- check_smoothness