+++ R/utils_pca_efa.R
@@ -77,7 +77,8 @@ predict.parameters_efa <- function(object, newdata = NULL, names = NULL, ...) {
 #' @export
 predict.parameters_pca <- predict.parameters_efa
 
-
+#' @export
+predict.parameters_clusters <- predict.parameters_efa
 
 
 
@@ -130,7 +131,6 @@ print.parameters_efa <- function(x, digits = 2, sort = FALSE, threshold = NULL, 
   }
 
 
-
   rotation_name <- attr(x, "rotation", exact = TRUE)
 
   if (rotation_name == "none") {
@@ -155,7 +155,6 @@ print.parameters_pca <- print.parameters_efa
 
 
 
-
 #' @keywords internal
 .text_components_variance <- function(x) {
   type <- attributes(x)$type
@@ -163,12 +162,19 @@ print.parameters_pca <- print.parameters_efa
     type <- "principal component"
   } else if (type %in% c("fa")) {
     type <- "latent factor"
+  } else if (type %in% c("kmeans")) {
+    type <- "cluster"
   } else {
     type <- paste0(type, " component")
   }
 
-
-  summary <- attributes(x)$summary
+  if(type == "cluster"){
+    summary <- as.data.frame(x)
+    variance <- attributes(x)$variance * 100
+  } else{
+    summary <- attributes(x)$summary
+    variance <- max(summary$Variance_Cumulative) * 100
+  }
 
   if (nrow(summary) == 1) {
     text <- paste0("The unique ", type)
@@ -177,18 +183,19 @@ print.parameters_pca <- print.parameters_efa
   }
 
   # rotation
-  if (attributes(x)$rotation != "none") {
+  if (!is.null(attributes(x)$rotation) && attributes(x)$rotation != "none") {
     text <- paste0(text, " (", attributes(x)$rotation, " rotation)")
   }
+
 
   text <- paste0(
     text,
     " accounted for ",
-    sprintf("%.2f", max(summary$Variance_Cumulative) * 100),
+    sprintf("%.2f", variance),
     "% of the total variance of the original data"
   )
 
-  if (nrow(summary) == 1) {
+  if (type == "cluster" || nrow(summary) == 1) {
     text <- paste0(text, ".")
   } else {
     text <- paste0(